Create stat submission form
This form allows players to automatically submit their stats to the site. It uses a three-step process and only submits on the final step. The previous two are displayed using JavaScript.

There is currently an extra allowance for missing SKIDs. This is to compensate for current entries that will not have one, and it will be removed once those entries are purged. After that point, SKIDs will be mandatory. All other fields are also mandatory.

All text is sanitized to only allow alphanumeric characters. Numbers must meet certain conditions based on the specific field. KDR is calculated, not input, and is checked for accuracy. KDR cannot be higher than 99.99 or lower than zero.

Add timer to end div to pause reload

Change oninput to addeventlistener for textboxes

Clear span warnings on successful KDR

Hide form when JavaScript disabled