Fix image bug
Image selection was clearing when cancel was hit, even if an image
was already selected. This behavior couldn't be changed, so I
just made it more obvious that a new image would have to be selected.

I moved checkImage to the front as the first function, and it will
now display an error message if files.length == 0. I also added
an image preview to the first screen as well as the second.